splits out monthly new chart from clients running total component
