+++ ui/app/components/clients/token/monthly-new.hbs
@@ -1,0 +1,55 @@
+<div class="chart-wrapper stacked-charts" data-test-running-total="monthly-charts">
+  <div class="single-chart-grid">
+    <div class="chart-header">
+      <h2 class="chart-title">Monthly new</h2>
+      <p class="chart-description">
+        Entity or non-entity clients which interacted with Vault for the first time during this date range, displayed per
+        month.
+      </p>
+    </div>
+
+    <div class={{concat (unless this.hasAverageNewClients "chart-empty-state ") "chart-container-wide"}}>
+      <Clients::VerticalBarChart
+        @dataset={{if this.hasAverageNewClients this.byMonthNewClients false}}
+        @chartLegend={{this.runningTotalLegend}}
+        @noDataTitle="No new clients"
+        @noDataMessage={{concat
+          "There is no new client data available for this "
+          (if @authMount "auth method" "namespace")
+          " in this date range"
+        }}
+      />
+    </div>
+
+    {{#if this.hasAverageNewClients}}
+      <div class="chart-subTitle has-top-padding-l" data-test-running-new-entity>
+        <h3 class="data-details">Average new entity clients per month</h3>
+        <p class="data-details">
+          {{format-number this.entityClientData.averageNewClients}}
+        </p>
+      </div>
+
+      <div class="data-details-top has-top-padding-l" data-test-running-new-nonentity>
+        <h3 class="data-details">Average new non-entity clients per month</h3>
+        <p class="data-details">
+          {{format-number this.nonEntityClientData.averageNewClients}}
+        </p>
+      </div>
+    {{/if}}
+
+    <div class="timestamp" data-test-running-total-timestamp>
+      {{#if @responseTimestamp}}
+        Updated
+        {{date-format @responseTimestamp "MMM d yyyy, h:mm:ss aaa" withTimeZone=true}}
+      {{/if}}
+    </div>
+
+    {{#if this.hasAverageNewClients}}
+      <div class="legend-right" data-test-running-total-legend>
+        {{#each this.runningTotalLegend as |legend idx|}}
+          <span class="legend-colors dot-{{idx}}"></span><span class="legend-label">{{capitalize legend.label}}</span>
+        {{/each}}
+      </div>
+    {{/if}}
+  </div>
+</div>
