+++ ui/app/styles/core/charts.scss
@@ -34,6 +34,9 @@
 .single-month-breakdown-nonentity {
   grid-column-start: 2;
 }
+.single-month-breakdown-sync {
+  grid-column-start: 3;
+}
 .stacked-charts {
   display: grid;
   width: 100%;
